+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### 🐞 Fixes
 
+- [#842](https://github.com/estruyf/vscode-front-matter/issues/842): Allow to set the `frontMatter.taxonomy.slugTemplate` setting to an empty string
 - [#845](https://github.com/estruyf/vscode-front-matter/issues/845): Fix empty values for number fields
 
 ## [10.3.0] - 2024-08-13 - [Release notes](https://beta.frontmatter.codes/updates/v10.3.0)